+++ examples/basic-effects.rs
@@ -238,23 +238,23 @@ impl EffectsRepository {
 
         let effects = vec![
             ("sweep in",
-                fx::sweep_in(Direction::LeftToRight, 30, screen_bg, (slow, QuadOut))),
-            ("sweep out/sweep in", sequence(&[
-                fx::sweep_out(Direction::DownToUp, 5, bg, (2000, QuadOut)),
-                fx::sweep_in(Direction::UpToDown, 5, bg, (2000, QuadOut)),
-                fx::sweep_out(Direction::UpToDown, 5, bg, (2000, QuadOut)),
-                fx::sweep_in(Direction::DownToUp, 5, bg, (2000, QuadOut)),
+                fx::sweep_in(Direction::LeftToRight, 30, 0, screen_bg, (slow, QuadOut))),
+            ("irregular sweep out/sweep in", sequence(&[
+                fx::sweep_out(Direction::DownToUp, 5, 20, bg, (2000, QuadOut)),
+                fx::sweep_in(Direction::UpToDown, 5, 20, bg, (2000, QuadOut)),
+                fx::sweep_out(Direction::UpToDown, 5, 20, bg, (2000, QuadOut)),
+                fx::sweep_in(Direction::DownToUp, 5, 20, bg, (2000, QuadOut)),
             ])),
             ("coalesce",
                 fx::coalesce((medium, CubicOut))),
             ("slide in/out", fx::repeating(sequence(&[
                 parallel(&[
                     fx::fade_from_fg(bg, (2000, ExpoInOut)),
-                    fx::slide_in(Direction::UpToDown, 20, Dark0Hard, medium),
+                    fx::slide_in(Direction::UpToDown, 20, 30, Dark0Hard, medium),
                 ]),
                 fx::sleep(medium),
                 fx::prolong_end(medium,
-                    fx::slide_out(Direction::LeftToRight, 80, Dark0Hard, medium),
+                    fx::slide_out(Direction::LeftToRight, 80, 30, Dark0Hard, medium),
                 ),
             ]))),
             ("change hue, saturation and lightness", sequence(&[